Refactored some stuff 2

+++ src/main/java/me/kaotich00/fwauctionhouse/objects/PendingSell.kt
@@ -2,4 +2,4 @@ package me.kaotich00.fwauctionhouse.objects
 
 import org.bukkit.inventory.ItemStack
 
-data class PendingSell(val listingId: Int, var itemStack: ItemStack, var buyerName: String, var totalCost: Float)
+data class PendingSell(val listingId: Int, var itemStack: ItemStack, var buyerName: String, var totalCost: Double)
